move processNode to ColumnStrategy and cleanup RowStrategy
